Eliminate race on SIGTERM #PSBM-78542
On SIGTERM we shutdown the screen, but in the meantime, there is
another thread handling events for the screen. There is a race
opportunity between listensocket check and socket processing
if shutdown is in progress. To eliminate the race we will close
the screen and sockets after worker thread exits.

Signed-off-by: Pavel Vokhmyanin <pvokhmyanin@virtuozzo.com>